move styles to separate file and extract component
