+++ Example/src/styles.ts
@@ -1,0 +1,19 @@
+import {StyleSheet} from 'react-native';
+
+const styles = StyleSheet.create({
+  title: {
+    fontSize: 36,
+    fontWeight: 'bold',
+  },
+  text: {
+    fontSize: 24,
+  },
+  scrollViewElement: {
+    alignItems: 'center',
+    margin: '1%',
+  },
+  scrollViewContainer: {
+    alignItems: 'center',
+  },
+});
+export default styles;
